file-organizer/requirements.txt
click>=8.1.0
pyyaml>=6.0
rich>=13.0.0
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
PyGithub>=2.0.0
python-dateutil>=2.8.0
